Linux bits
Implement missing pieces for Linux support:
- CMake buildsystem
- posix/linux specific file- and pathhandling bits.
  this implementation also takes case-sensitivity of
  unix filesystems into account.
- add infrastructure to look up system files (shaders, doc,
  fonts, ...) independently from the main binary (for distro
  packaging).
- fix up the linux thread implementation.
- fix up 32bit target build.
- crash handler skeleton.
- can look for game data files in local Steam installation on
  multiple libraries.
  (GOG linux games do not leave trails about their installation
   paths in any central location, so gog detection is missing).

based partly on previous work by Matthew Gilmore.

+++ TheForceEngine/TFE_DarkForces/weapon.cpp
@@ -923,7 +923,7 @@ namespace TFE_DarkForces
 			}
 			else
 			{
-				s32 elapsedFrames = min(s_weaponAnimState.frameCount, elapsed / s_weaponAnimState.ticksPerFrame);
+				s32 elapsedFrames = min((s32)s_weaponAnimState.frameCount, (s32)(elapsed / s_weaponAnimState.ticksPerFrame));
 				s_curPlayerWeapon->xOffset += s_weaponAnimState.xSpeed * elapsedFrames;
 				s_curPlayerWeapon->yOffset += s_weaponAnimState.ySpeed * elapsedFrames;
 				s_weaponAnimState.frameCount -= elapsedFrames;
