+++ TheForceEngine/TFE_Jedi/Level/robjData.cpp
@@ -191,7 +191,7 @@ namespace TFE_Jedi
 			SERIALIZE(ObjState_InitVersion, writeCount, 0);
 			if (!s_objData.objectList)
 			{
-				const u32 initChunkCount = max(1, (writeCount + 255) >> 8);
+				const u32 initChunkCount = (u32)(s32)max((s32)1, (s32)((writeCount + 255) >> 8));
 				s_objData.objectList = TFE_Memory::createChunkedArray(sizeof(SecObject), 256, initChunkCount, s_levelRegion);
 			}
 			else
